bugfix(windows): Use real dimensions of a screen to support scaling
On windows, when the `EnumDisplayMonitors` function is used with screen
scaling, windows treats the application as DPI unaware leading to the
dimensions returned in the callback parameters not reflecting the actual
resolution of the display.

This leads to a screenshot of the scaled display only containing a
portion of the visible image. To combat this, a new method
`getMonitorRealSize` has been introduced that uses the monitor handle
from the callback parameters to obtain the display adapter name.

With the display adapter name, device configuration can be obtained that
contains the real screen resolution and position, these are then
returned and used as the bounds of the display.